models/DeepConvNet.py
# models/DeepConvNet.py
import torch.nn as nn

class DeepConvNet(nn.Module):
    def __init__(self, num_classes, in_channels, input_length):
        super().__init__()
        self.conv1 = nn.Conv2d(1, 25, kernel_size=(1, 5))
        self.conv2 = nn.Conv2d(25, 50, kernel_size=(in_channels, 1))
        self.bn2 = nn.BatchNorm2d(50)
        self.pool2 = nn.MaxPool2d(kernel_size=(1, 2))
        # 后续卷积层...
        self.fc = nn.Linear(50*10, num_classes)  # 根据输入长度调整

    def forward(self, x):
        # 实现前向传播逻辑
        return x